tracing: pass call callbacks on correct provider
This will make it so callbacks are only called on the correct provider,
but still reserve the async stack.

Also the provider type (e.g. TCP) will be passed as the second argument
to the create() callback.

process.nextTick() is a special case. Because of its implicit call style
within core code the callbacks will always be called for all nextTick().